removed sign comparison warnings. only affects tests

+++ src/test/mcmc/hmc/hamiltonians/diag_e_metric_test.cpp
@@ -71,7 +71,7 @@ TEST(McmcDiagEMetric, gradients) {
   
   Eigen::VectorXd g1 = metric.dtau_dq(z);
   
-  for (int i = 0; i < z.q.size(); ++i) {
+  for (size_t i = 0; i < z.q.size(); ++i) {
     
     double delta = 0;
     
@@ -94,7 +94,7 @@ TEST(McmcDiagEMetric, gradients) {
   
   Eigen::VectorXd g2 = metric.dtau_dp(z);
   
-  for (int i = 0; i < z.q.size(); ++i) {
+  for (size_t i = 0; i < z.q.size(); ++i) {
     
     double delta = 0;
     
@@ -114,7 +114,7 @@ TEST(McmcDiagEMetric, gradients) {
   
   Eigen::VectorXd g3 = metric.dphi_dq(z);
   
-  for (int i = 0; i < z.q.size(); ++i) {
+  for (size_t i = 0; i < z.q.size(); ++i) {
     
     double delta = 0;
